chore: Split e2e by type (#3283)
