+++ tests/README.md
@@ -9,9 +9,9 @@
 ### All tests
 
 ```bash
-go test -v setup_test.go        # Only needs to be run once.
+go test -v utils/setup_test.go        # Only needs to be run once.
 go test -v ./scalers_go/...
-go test -v cleanup_test.go      # Skip if you want to keep testing.
+go test -v utils/cleanup_test.go      # Skip if you want to keep testing.
 ```
 
 ### Specific test
@@ -26,21 +26,26 @@ Refer to [this](https://pkg.go.dev/testing) for more information about testing i
 
 The test script will run in 3 phases:
 
-- **Setup:** This is done in [`setup_test.go`](setup_test.go). If you're adding any tests to the KEDA install / setup process, you need to add it to this file. `setup_test.go` deploys KEDA to the `keda` namespace, updates the image to
+- **Setup:** This is done in [`utils/setup_test.go`](utils/setup_test.go). If you're adding any tests to the KEDA install / setup process, you need to add it to this file. `utils/setup_test.go` deploys KEDA to the `keda` namespace, updates the image to
 `kedacore/keda:main`.
 
-    After `setup_test.go` is done, we expect to have KEDA setup in the `keda` namespace.
+    After `utils/setup_test.go` is done, we expect to have KEDA setup in the `keda` namespace.
 
 - **Tests:** Currently there are only scaler tests in `tests/scalers_go/`. Each test is kept in its own package. This is to prevent conflicting variable declarations for commoly used variables (**ex -** `testNamespace`). Individual scaler tests are run
 in parallel, but tests within a file can be run in parallel or in series. More about tests below.
 
-- **Global cleanup:** This is done in [`cleanup_test.go`](cleanup_test.go). It cleans up all the resources created in `setup_test.go`.
+- **Global cleanup:** This is done in [`utils/cleanup_test.go`](utils/cleanup_test.go). It cleans up all the resources created in `utils/setup_test.go`.
 
 ## Adding tests
 
 - Tests are written using `Go`'s default [`testing`](https://pkg.go.dev/testing) framework, and [`testify`](https://pkg.go.dev/github.com/stretchr/testify).
-- Each scaler should be in its own package, **ex -** `scalers_go/azure_queue/azure_queue_test.go`, or `scalers_go/kafka/kafka_test.go`, etc
+- Each e2e test should be in its own package, **ex -** `scalers_go/azure_queue/azure_queue_test.go`, or `scalers_go/kafka/kafka_test.go`, etc
 - Each test file is expected to do its own setup and clean for resources.
+
+Test are split in different folders based on what it's testing:
+- `internals`: KEDA internals (ie: HPA related stuff).
+- `scalers_go`: Anything related with scalers.
+- `secret-providers`: Anything related with how KEDA gets the secrets for working (ie: pod-identity, vault, etc).
 
 #### ⚠⚠ Important: ⚠⚠
 >
