+++ tests/secret-providers/azure_keyvault/azure_keyvault_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package azure_keyvault_queue_test
+package azure_keyvault_test
 
 import (
 	"context"
