+++ tests/secret-providers/azure_workload_identity/azure_workload_identity_test.go
