+++ tests/secret-providers/azure_workload_identity_user_assigned/azure_workload_identity_user_assigned_test.go
